Move input to command package
